unit test the reorder helper; fixed lint error
