+++ src/js/components/settings/table/SettingsTable.jsx
@@ -7,6 +7,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { startCase } from 'lodash';
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { reorder } from 'helpers/dndHelper';
 import { DragDropContext, Droppable, Draggable } from 'react-beautiful-dnd';
 import ImpactDropdown from './ImpactDropdown';
 
@@ -22,19 +23,11 @@ const defaultProps = {
 
 const columns = ['significance', 'impact', 'description'];
 
-// a function to help with reordering the result
-export const reorder = (list, startIndex, endIndex) => {
-    const result = Array.from(list);
-    const [removed] = result.splice(startIndex, 1);
-    result.splice(endIndex, 0, removed);
-    return result;
-};
-
 const getItemStyle = (isDragging, draggableStyle) => ({
-    userSelect: "none",
+    userSelect: 'none',
     // change background color if dragging
-    background: isDragging ? "#9bdaf1" : "",
-    boxShadow: isDragging ? "0 2px 5px rgba(0, 0, 0, 0.15)" : "0 0",
+    background: isDragging ? '#9bdaf1' : '',
+    boxShadow: isDragging ? '0 2px 5px rgba(0, 0, 0, 0.15)' : '0 0',
     // styles we need to apply on draggables
     ...draggableStyle
 });
@@ -75,6 +68,7 @@ export default class SettingsTable extends React.Component {
                             <tbody className="broker-table__body">
                                 {this.props.results.map((row, index) => (
                                     <Draggable key={row.label} draggableId={row.label} index={index}>
+                                        {/* eslint-disable-next-line no-shadow */}
                                         {(provided, snapshot) => (
                                             <tr
                                                 className="settings-table__row"
