+++ src/js/helpers/dndHelper.js
@@ -1,0 +1,8 @@
+/* eslint-disable import/prefer-default-export */
+export const reorder = (list, startIndex, endIndex) => {
+    const result = Array.from(list);
+    const [removed] = result.splice(startIndex, 1);
+    result.splice(endIndex, 0, removed);
+    return result;
+};
+/* eslint-enable import/prefer-default-export */
